ppxx/ecommerce/requirements.txt
djangorestframework
djangorestframework-simplejwt
Pillow